ui: refactored i9n download filename to being opstrace-<tenant>-integration-<integration.kind>.yaml. also DRYd up some dup code
