+++ packages/app/src/client/viewsBasic/integrationDefs/common/ViewConfigButtonModal.tsx
@@ -1,0 +1,73 @@
+/**
+ * Copyright 2021 Opstrace, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import React from "react";
+
+import { Button } from "client/components/Button";
+import { makeStyles, Theme, createStyles } from "@material-ui/core/styles";
+import Modal from "@material-ui/core/Modal";
+
+const useStyles = makeStyles((theme: Theme) =>
+  createStyles({
+    modal: {
+      display: "flex",
+      justifyItems: "center",
+      alignItems: "center",
+      justifyContent: "center",
+      height: "500px",
+      overflow: "scroll"
+    },
+    paper: {
+      backgroundColor: theme.palette.background.paper,
+      border: "2px solid #000",
+      boxShadow: theme.shadows[5],
+      padding: theme.spacing(2, 4, 3)
+    }
+  })
+);
+
+export const ViewConfigButtonModal = ({
+  filename,
+  config
+}: {
+  filename: string;
+  config: string;
+}) => {
+  const classes = useStyles();
+  const [open, setOpen] = React.useState(false);
+
+  const handleOpen = () => {
+    setOpen(true);
+  };
+
+  const handleClose = () => {
+    setOpen(false);
+  };
+
+  return (
+    <>
+      <Button variant="contained" size="small" onClick={handleOpen}>
+        View YAML
+      </Button>
+      <Modal open={open} onClose={handleClose} className={classes.modal}>
+        <div className={classes.paper}>
+          <h2>{filename}</h2>
+          <pre>{config}</pre>
+        </div>
+      </Modal>
+    </>
+  );
+};
